The most significant changes involve the addition of new variables and methods to control the behavior of enemy ships in the game, the introduction of a debug mode for easier testing and debugging, and the implementation of a custom hash function for efficient lookup in a hashmap.
1. New variables `firingDistance`, `turningSpeed`, and `enemySpeedMultiplier` were added to the `EnemyShipHandler_PG.h` file to control the behavior of enemy ships. Setter methods for these variables were added in `EnemyShipInputHandler_PG.h` and `EnemyShipMovementHandler_PG.h`, and their default values were set to 0. The `addEnemyShips` method in `EnemyShipHandler_PG.cpp` was updated to set these new variables for each enemy ship [1,2,3,10].

2. A debug mode was added to the `World` class in `World_PG.h` for easier testing and debugging. New variables were added to control the number of land masses and enemy ships, and their distances, which can be set differently for debug and non-debug modes. The `setUpWorld` and `createWorld` methods in `World_PG.cpp` were updated to use these new variables. A new `pirateGameDebug` function was added in `main.cpp` to start the game in debug mode, and an option to start the game in debug mode was added in the `main` function in `main.cpp` [8,9,18,19,20].

3. A custom hash function for `std::pair<int, int>` was added in `ShipHashmap_PG.h` for efficient lookup in a hashmap where the keys are pairs of integers. The `addEnemyShip` and `removeEnemyShip` methods in `ShipHashmap_PG.cpp` were updated to use this new custom hash function [6,14,15].

Other changes include the increase of the cell size in the `GlobalMap_PG.h` file from 10 to 100, affecting the granularity of the game map [5], the fix of the file path for loading the Times New Roman font in `GlobalFontHandler_PG.h` [4], and the renaming of the `update` method in `ShipHashmap_PG.h` to `updateEnemyShipPosition` for better descriptiveness [7]. The `updateEnemyShipPosition` method in `ShipHashmap_PG.cpp` was updated for more efficient position updating [16]. The `handleCannonFire` method in `EnemyShipInputHandler_PG.cpp` and the `setSpriteRotation` method in `EnemyShipMovementHandler_PG.cpp` were updated to use a static cast for floating point arithmetic [11,13].

+++ PirateGame/Resource/EnemyShipInputHandler_PG.cpp
@@ -4,7 +4,7 @@ using namespace PirateGame;
 
 void EnemyShipInputHandler::handleCannonFire() {
 	// Only fire if the player is within firingRange
-	float distance = sqrt(pow(playerPos.x - sprite.getPosition().x, 2) + pow(playerPos.y - sprite.getPosition().y, 2));
+	float distance = static_cast<float>(sqrt(pow(playerPos.x - sprite.getPosition().x, 2) + pow(playerPos.y - sprite.getPosition().y, 2)));
 	if (distance > firingDistance) {
 		return;
 	}
